requirements.txt
Django
Pillow
django-colorfield
